requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pydantic-settings